question2_sql_analysis/requirements.txt
mysql-connector-python>=8.2.0
